Changed the way wall collisions are handled.
Also fixed the tests. The Wall class is now abstract and got new
abstract methods. Also did some code review and changed the way the
score is added. This is not done via Level anymore but directly by
calling the normaldriver in the observer (the way score is added should
change right?..) Also gave bubbles their own score worth and used that
to add score.
<review,testing,organisation,documentation,logic,interface